W4 flags

+++ main/plot_models/column_plot.cpp
@@ -13,7 +13,7 @@ void ColumnPlot::Draw() {
 
     QVector<double> xAxis_data;
     QVector<double> yAxis_data;
-    for (int j = 0; j < variable.GetMeasurementsCount(); j++) {
+    for (size_t j = 0; j < variable.GetMeasurementsCount(); j++) {
       if (variable.measurements[j]) {
         xAxis_data.push_back((j + 0.55 + i * (0.9 / n) + 0.9 / (2 * n)));
         yAxis_data.push_back(variable.measurements[j]);
